pyproject.toml
[project]
name = "mcp-doc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "httpx>=0.28.1",
    "langchain-community>=0.3.20",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.3.9",
    "langgraph>=0.3.18",
    "lxml>=5.3.1",
    "markdownify>=1.1.0",
    "matplotlib>=3.10.1",
    "mcp[cli]>=1.4.1",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pydantic>=1.10.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-socket>=0.7.0",
    "pytest-timeout>=2.3.1",
    "pyyaml>=6.0.1",
    "ruff>=0.9.7",
    "scikit-learn>=1.6.1",
    "transformers>=4.50.0",
]
